+++ packages/validators/src/withMessages/and.js
@@ -1,4 +1,5 @@
 import and from '../raw/and'
+import { withMessage, withParams } from '../common'
 
 /**
  * Validate if all validators match.
@@ -6,8 +7,9 @@ import and from '../raw/and'
  * @returns {NormalizedValidator}
  */
 export default function (...validators) {
-  return {
-    $validator: and(...validators),
-    $message: 'The value does not match all of the provided validators'
-  }
+  return withParams({ type: 'and' },
+    withMessage('The value does not match all of the provided validators',
+      and(...validators)
+    )
+  )
 }