+++ packages/validators/src/withMessages/decimal.js
@@ -6,5 +6,6 @@ import decimal from '../raw/decimal'
  */
 export default {
   $validator: decimal,
-  $message: 'Value must be decimal'
+  $message: 'Value must be decimal',
+  $params: { type: 'decimal' }
 }