+++ packages/validators/src/withMessages/email.js
@@ -6,5 +6,6 @@ import email from '../raw/email'
  */
 export default {
   $validator: email,
-  $message: 'Value is not a valid email address'
+  $message: 'Value is not a valid email address',
+  $params: { type: 'email' }
 }