+++ packages/validators/src/withMessages/integer.js
@@ -1,10 +1,11 @@
 import integer from '../raw/integer'
 
-  /**
-   * Validate if value is integer.
-   * @type {NormalizedValidator}
-   */
+/**
+ * Validate if value is integer.
+ * @type {NormalizedValidator}
+ */
 export default {
   $validator: integer,
-  $message: 'Value is not an integer'
+  $message: 'Value is not an integer',
+  $params: { type: 'integer' }
 }