+++ packages/validators/src/withMessages/requiredUnless.js
@@ -5,7 +5,10 @@ import requiredUnless from '../raw/requiredUnless'
  * @param {Boolean | String | function(): (Boolean | Promise<boolean>)} prop
  * @return {NormalizedValidator}
  */
-export default prop => ({
-  $validator: requiredUnless(prop),
-  $message: 'The value is required'
-})
+export default function (prop) {
+  return {
+    $validator: requiredUnless(prop),
+    $message: 'The value is required',
+    $params: { type: 'requiredUnless', prop }
+  }
+}