+++ packages/vuelidate/src/core.js
@@ -37,7 +37,7 @@ let ROOT_PATH = '__root'
 /**
  * Sorts the validators for a state tree branch
  * @param {Object<NormalizedValidator|Function>} validationsRaw
- * @return {{ rules: Object<NormalizedValidator>, nestedValidators: Object, config: Object }}
+ * @return {{ rules: Object<NormalizedValidator>, nestedValidators: Object, config: GlobalConfig }}
  */
 function sortValidations (validationsRaw = {}) {
   const validations = unwrap(validationsRaw)
@@ -104,7 +104,8 @@ function normalizeValidatorResponse (result) {
  * @param {Ref<*>} model
  * @param {Ref<Boolean>} $pending
  * @param {Ref<Boolean>} $dirty
- * @param {Object} config
+ * @param {GlobalConfig} config
+ * @param {boolean} config.$lazy
  * @param {Ref<*>} $response
  * @param {VueInstance} instance
  * @param {Ref<*>[]} watchTargets
@@ -157,7 +158,7 @@ function createAsyncResult (rule, model, $pending, $dirty, { $lazy }, $response,
  * @param {Validator} rule
  * @param {Ref<*>} model
  * @param {Ref<Boolean>} $dirty
- * @param {Object} config
+ * @param {GlobalConfig} config
  * @param {Boolean} config.$lazy
  * @param {Ref<*>} $response
  * @param {VueInstance} instance
@@ -185,7 +186,7 @@ function createSyncResult (rule, model, $dirty, { $lazy }, $response, instance) 
  * @param {NormalizedValidator} rule
  * @param {Ref<*>} model
  * @param {Ref<boolean>} $dirty
- * @param {Object} config
+ * @param {GlobalConfig} config
  * @param {VueInstance} instance
  * @return {{ $params: *, $message: Ref<String>, $pending: Ref<Boolean>, $invalid: Ref<Boolean>, $response: Ref<*>, $unwatch: WatchStopHandle }}
  */
@@ -277,11 +278,12 @@ function createValidatorResult (rule, model, $dirty, config, instance, validator
  * @param {String} key - Key for the current state tree
  * @param {ResultsStorage} [resultsCache] - A cache map of all the validators
  * @param {String} [path] - the current property path
- * @param {Object} [config] - the config object
+ * @param {GlobalConfig} [config] - the config object
  * @param {VueInstance} instance
+ * @param {ComputedRef<Object>} externalResults
  * @return {ValidationResult | {}}
  */
-function createValidationResults (rules, model, key, resultsCache, path, config, instance) {
+function createValidationResults (rules, model, key, resultsCache, path, config, instance, externalResults) {
   // collect the property keys
   const ruleKeys = Object.keys(rules)
 
@@ -329,7 +331,22 @@ function createValidationResults (rules, model, key, resultsCache, path, config,
     )
   })
 
+  result.$externalResults = computed(() => {
+    if (!externalResults.value) return []
+    return [].concat(externalResults.value).map((stringError, index) => ({
+      $propertyPath: path,
+      $property: key,
+      $validator: '$externalResults',
+      $uid: `${path}-${index}`,
+      $message: stringError,
+      $params: {},
+      $response: null,
+      $pending: false
+    }))
+  })
+
   result.$invalid = computed(() =>
+    !!result.$externalResults.value.length ||
     ruleKeys.some(ruleKey => unwrap(result[ruleKey].$invalid))
   )
 
@@ -356,6 +373,7 @@ function createValidationResults (rules, model, key, resultsCache, path, config,
         $pending: res.$pending
       })
     })
+    .concat(result.$externalResults.value)
   )
 
   result.$errors = computed(() => result.$dirty.value
@@ -378,10 +396,12 @@ function createValidationResults (rules, model, key, resultsCache, path, config,
  * @param {Object} nestedState - Current state
  * @param {String} path - Path to current property
  * @param {ResultsStorage} resultsCache - Validations cache map
- * @param {Object} config - The config object
+ * @param {GlobalConfig} config - The config object
+ * @param {VueInstance} instance - The current Vue instance
+ * @param {ComputedRef<object>} nestedExternalResults - The external results for this nested collection
  * @return {{}}
  */
-function collectNestedValidationResults (validations, nestedState, path, resultsCache, config, instance) {
+function collectNestedValidationResults (validations, nestedState, path, resultsCache, config, instance, nestedExternalResults) {
   const nestedValidationKeys = Object.keys(validations)
 
   // if we have no state, return empty object
@@ -396,7 +416,8 @@ function collectNestedValidationResults (validations, nestedState, path, results
       parentKey: path,
       resultsCache,
       globalConfig: config,
-      instance
+      instance,
+      externalResults: nestedExternalResults
     })
     return results
   }, {})
@@ -405,8 +426,8 @@ function collectNestedValidationResults (validations, nestedState, path, results
 /**
  * Generates the Meta fields from the results
  * @param {ValidationResult|{}} results
- * @param {Object<ValidationResult>[]} nestedResults
- * @param {Object<ValidationResult>[]} childResults
+ * @param {Object.<string, ValidationResult>[]} nestedResults
+ * @param {Object.<string, ValidationResult>[]} childResults
  * @return {{$anyDirty: Ref<Boolean>, $error: Ref<Boolean>, $invalid: Ref<Boolean>, $errors: Ref<ErrorObject[]>, $dirty: Ref<Boolean>, $touch: Function, $reset: Function }}
  */
 function createMetaFields (results, nestedResults, childResults) {
@@ -546,8 +567,10 @@ function createMetaFields (results, nestedResults, childResults) {
  * @param {String} [params.key] - Current state property key. Used when being called on nested items
  * @param {String} [params.parentKey] - Parent state property key. Used when being called recursively
  * @param {Object<ValidationResult>} [params.childResults] - Used to collect child results.
- * @param {ResultsStorage} resultsCache - The cached validation results
- * @param {VueInstance} instance - The current Vue instance
+ * @param {ResultsStorage} params.resultsCache - The cached validation results
+ * @param {VueInstance} params.instance - The current Vue instance
+ * @param {GlobalConfig} params.globalConfig - The validation config, passed to this setValidations instance.
+ * @param {Reactive<object> | Ref<Object>} params.externalResults - External validation results
  * @return {UnwrapNestedRefs<VuelidateState>}
  */
 export function setValidations ({
@@ -558,7 +581,8 @@ export function setValidations ({
   childResults,
   resultsCache,
   globalConfig = {},
-  instance
+  instance,
+  externalResults
 }) {
   const path = parentKey ? `${parentKey}.${key}` : key
 
@@ -578,11 +602,20 @@ export function setValidations ({
     })
     : state
 
+  // cache the external results, so we can revert back to them
+  const cachedExternalResults = { ...(unwrap(externalResults) || {}) }
+
+  const nestedExternalResults = computed(() => {
+    const results = unwrap(externalResults)
+    if (!key) return results
+    return results ? unwrap(results[key]) : undefined
+  })
+
   // Use rules for the current state fragment and validate it
-  const results = createValidationResults(rules, nestedState, key, resultsCache, path, mergedConfig, instance)
+  const results = createValidationResults(rules, nestedState, key, resultsCache, path, mergedConfig, instance, nestedExternalResults)
   // Use nested keys to repeat the process
   // *WARN*: This is recursive
-  const nestedResults = collectNestedValidationResults(nestedValidators, nestedState, path, resultsCache, mergedConfig, instance)
+  const nestedResults = collectNestedValidationResults(nestedValidators, nestedState, path, resultsCache, mergedConfig, instance, nestedExternalResults)
 
   // Collect and merge this level validation results
   // with all nested validation results
@@ -652,6 +685,22 @@ export function setValidations ({
     return (childResults.value || {})[key]
   }
 
+  function $clearExternalResults () {
+    if (isRef(externalResults)) {
+      externalResults.value = cachedExternalResults
+    } else {
+      // if the external results state was empty, we need to delete every property, one by one
+      if (Object.keys(cachedExternalResults).length === 0) {
+        Object.keys(externalResults).forEach((k) => {
+          delete externalResults[k]
+        })
+      } else {
+        // state was not empty, so we just assign it back into the current state
+        Object.assign(externalResults, cachedExternalResults)
+      }
+    }
+  }
+
   return reactive({
     ...results,
     // NOTE: The order here is very important, since we want to override
@@ -671,7 +720,8 @@ export function setValidations ({
     // if there are no child results, we are inside a nested property
     ...(childResults && {
       $getResultsForChild,
-      $validate
+      $validate,
+      $clearExternalResults
     }),
     // add each nested property's state
     ...nestedResults